% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBH_recon.R
\name{DBH_recon}
\alias{DBH_recon}
\title{Input tree ring and ground joined file}
\usage{
DBH_recon(treering_ground_join)
}
\arguments{
\item{treering_ground_join}{inputs from tree_ring_ground function results}
}
\value{
A file a merged ground data and tree ring data reconstructed as DBH
}
\description{
This function loads a path file of EXCEL format xlsx field metadata. Field metadata needs to have columns:
Azimuth, Distance, Species, Height, DBH and TreeID in the exact format
Rows are the year increments
}
