% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_parse.R
\name{ground_parse}
\alias{ground_parse}
\title{Load a EXCEL ground metadata file}
\usage{
ground_parse(coor, ground)
}
\arguments{
\item{coor}{c(x,y) coordinate of plot center}

\item{ground}{Path to ground metadata}
}
\value{
A file with new lat an lon coordinates of each tree based on distance and Azimuth
}
\description{
This function loads a path file of EXCEL format xlsx field metadata. Field metadata needs to have columns:
Azimuth, Distance, Species, Height, DBH and TreeID in the exact format
}
