% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treering_ground_join.R
\name{treering_ground}
\alias{treering_ground}
\title{Load treering_parse and ground_parse file}
\usage{
treering_ground(treering_parse, ground_parse, plot)
}
\arguments{
\item{treering_parse}{file results from treering_parse}

\item{ground_parse}{file results from ground_parse}

\item{plot}{user adds "PLOTID" to subset e.g. "HF" typically a two letter identifier}
}
\value{
A joined summarized treering file and ground metadata
}
\description{
Function creates a mean summary of all cores for a single tree and merges with ground data.
User must select plot if more than one plot in file.
}
