% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/treering_parse.R
\name{treering_parse}
\alias{treering_parse}
\title{Load a RWL file}
\usage{
treering_parse(tree)
}
\arguments{
\item{tree}{Path to the input file}
}
\value{
A parsed, transposed file with added name columns for tree, plot and core identifier
}
\description{
This function loads a path file of rwl format. rwl column names are 'XX00YY'XX as plot, 00 tree, YY core.
Rows are the year increments
}
